Bulk Update
Lots of updates were made today.
1. Import Eigen as submodule again.  - But if you want it lighter, you can still see the Eigen folder in the old commit.
2. Re-configure the .sln and .json file. - Now you can easily build a Release version.
3. Reduce computation by retaining process values. - Mostly by noticing the variables that are the same during iteration so we don't need to calculate it so many times.

+++ Eigen/unsupported/Eigen/src/MatrixFunctions/MatrixPower.h
@@ -10,6 +10,9 @@
 #ifndef EIGEN_MATRIX_POWER
 #define EIGEN_MATRIX_POWER
 
+// IWYU pragma: private
+#include "./InternalHeaderCheck.h"
+
 namespace Eigen {
 
 template<typename MatrixType> class MatrixPower;
@@ -281,7 +284,7 @@ inline int MatrixPowerAtomic<MatrixType>::getPadeDegree(long double normIminusT)
 #endif
   int degree = 3;
   for (; degree <= maxPadeDegree; ++degree)
-    if (normIminusT <= maxNormForPade[degree - 3])
+    if (normIminusT <= static_cast<long double>(maxNormForPade[degree - 3]))
       break;
   return degree;
 }
